inventory reworking skeleton
