+++ Inventory.txt
@@ -1,0 +1,9 @@
+inventory slots drawn based on the indexes of the items array.
+
+All the drag positions x and y for the items and text, that data can be given to the display.
+
+
+equipItem function in inventory, since its currently in the inventory (items array)
+unequipItem function in equipment, since its currently in the equipment (equip array).
+The only player aspect of a method like usePotion is incrementing the players HP. Therefore, the InventoryData will need a reference to the player.
+
