+++ js/Maps.js
@@ -1,4 +1,6 @@
+import UserInput from './UserInput.js';
 import Player from "./Player.js";
+import InventoryData from "./InventoryData.js";
 import {Monster, MonsterManager} from "./Monsters.js";
 
 export default class Map extends Phaser.Scene {
@@ -27,16 +29,19 @@ export default class Map extends Phaser.Scene {
         this.matter.world.convertTilemapLayer(background);
         this.matter.world.convertTilemapLayer(environment);
 
-        this.player = new Player(this, 320, 320, this.gameState, this.inventoryDisplay);
+        this.handleWheel = this.handleWheel.bind(this);
+        this.userInput = new UserInput(this, { onWheel: this.handleWheel });
+        
+        this.player = new Player(this, 320, 320, this.gameState);
         this.gameState.loadPlayerState(this.player);
+        this.scene.launch('PlayerStatusBars', { player: this.player });
+
+        this.inventory = new InventoryData(this, this.gameState)
+        this.gameState.loadInventoryState(this.inventory)
+        this.scene.launch('InventoryDisplay', { inventory: this.inventory});
 
         this.monsterManager = new MonsterManager(this, this.player);
         this.spawnMonster();
-
-        this.inventoryData = this.gameState.inventoryData;
-        this.scene.launch('InventoryDisplay', { gameState: this.gameState, player: this.player });
-
-        this.scene.launch('PlayerStatusBars', { player: this.player });
 
         let camera = this.cameras.main;
         camera.zoom = 1.4;
@@ -44,16 +49,16 @@ export default class Map extends Phaser.Scene {
         camera.setLerp(0.1,0.1);
         camera.setBounds(0, 0, this.game.config.width,this.game.config.height);
 
-        this.input.on('wheel', (pointer, gameObjects, deltaX, deltaY, deltaZ) => {
-            const zoomChange = 0.1;
-            if (deltaY > 0) {
-                this.cameras.main.zoom = Math.max(this.cameras.main.zoom - zoomChange, 1);
-            } else {
-                this.cameras.main.zoom = Math.min(this.cameras.main.zoom + zoomChange, 3.0);
-            }
-        });
-
         this.events.once('shutdown', this.shutdown, this);
+    }
+
+    handleWheel(deltaY) {
+        const zoomChange = 0.1;
+        if (deltaY > 0) {
+            this.cameras.main.zoom = Math.max(this.cameras.main.zoom - zoomChange, 1);
+        } else {
+            this.cameras.main.zoom = Math.min(this.cameras.main.zoom + zoomChange, 3.0);
+        }
     }
     
     spawnMonster() {}
@@ -86,6 +91,7 @@ export class Map1 extends Map {
         super.update();
         if (this.player.sprite.x > this.sys.game.config.width) {
           this.gameState.savePlayerState(this.player);
+          this.gameState.saveInventoryState(this.inventory);
           this.gameState.playerPosition.x = 0;
           this.scene.start('Map2');
         }
@@ -108,6 +114,7 @@ export class Map2 extends Map {
         super.update();
         if (this.player.sprite.x < 0) {
           this.gameState.savePlayerState(this.player);
+          this.gameState.saveInventoryState(this.inventory);
           this.gameState.playerPosition.x = this.sys.game.config.width;
           this.scene.start('Map1');
         }
