+++ js/Player.js
@@ -8,10 +8,9 @@ export default class Player {
     scene.load.animation('hero_anims', 'assets/images/hero_anims.json');
   }
 
-  constructor(scene, x, y, gameState, inventoryDisplay) {
+  constructor(scene, x, y, gameState) {
     this.scene = scene;
     this.gameState = gameState;
-    this.inventoryDisplay = inventoryDisplay;
     this.gameState.setPlayerXP(0);
     this.gameState.setPlayerMaxXP(50);
     this.gameState.setPlayerLevel(1);
@@ -59,23 +58,6 @@ export default class Player {
     this.scene.matter.world.on('collisionactive', this.handleCollision, this);
     this.manaRegeneration();
   }
-  
-  setInventoryDisplay(inventoryDisplay) {
-    this.inventoryDisplay = inventoryDisplay;
-}
-
-  usePotion() {
-    let potionIndex = this.gameState.inventoryData.items.findIndex(item => item && item.name === 'potion');
-    if (potionIndex !== -1) {
-        this.gameState.inventoryData.decrementQuantity(potionIndex);
-        this.gameState.setPlayerHP(this.gameState.getPlayerHP() + 50);
-        this.inventoryDisplay.updateItemQuantity(potionIndex);
-        if (this.gameState.inventoryData.items[potionIndex] === null) {
-          this.inventoryDisplay.destroyItem(potionIndex);
-      }
-        console.log(`Player Used Potion. Current HP ${this.gameState.getPlayerHP()}`)
-    }
-}
 
   gainXP(monster) {
     const currentXP = this.gameState.getPlayerXP();
