+++ js/UserInput.js
@@ -4,10 +4,10 @@ Handles all things associated with user input:
 --> Clicking Things (Changing volume, changing options, click load/save buttons)
 --> Entering Information (inputting values in options or text box prompts in game)
 */
-
 export default class UserInput {
-  constructor(scene) {
+  constructor(scene, handlers) {
     this.scene = scene;
+
     this.cursors = this.scene.input.keyboard.addKeys({
       up: Phaser.Input.Keyboard.KeyCodes.W,
       down: Phaser.Input.Keyboard.KeyCodes.S,
@@ -24,5 +24,13 @@ export default class UserInput {
       arrowLeft: Phaser.Input.Keyboard.KeyCodes.LEFT,
       arrowRight: Phaser.Input.Keyboard.KeyCodes.RIGHT,
     });
+
+    this.scene.input.on('wheel', handlers.onWheel);
+    this.scene.input.on('pointerover', handlers.onPointerOver);
+    this.scene.input.on('pointerout', handlers.onPointerOut);
+    this.scene.input.on('pointerdown', handlers.onPointerDown);
+    this.scene.input.on('drag', handlers.onDrag);
+    this.scene.input.on('dragstart', handlers.onDragStart);
+    this.scene.input.on('dragend', handlers.onDragEnd);
   }
 }
